requirements.txt
streamlit
groq
unbabel-comet